Allow overriding remote settings per environment.
Allow overriding remote settings per environment in the properties.yml
file. Itmakes sense for some settings to differ between environments.
For example the `cleandir_limit` setting. It makes sense to maintain
only one or two releases on staging to minimize storage, but to maintain
three or even more releases on production to be able to rollback when
things go wrong.