Reverse the symlink order after loading config
Follow-up of the issue we tried to fix in [098d1a7]. We still had an
issue where a cached twig template file was stored in the
sites/default/files directory of a Drupal site before the symlink was
created. So even though the command to remove the directory and the
command to create the symlink followed each other directly, there were
still cases where that directory was recreated to store the compiled
twig file.

In this commit we reverse the order of the symlinks, so that the
`current` symlink is the last (in stead of the first) one to be created.
This makes sure that no traffic is handled by the new release before all
other symlinks (including sites/default/files and
sites/default/settings.php) are created. This _should_ make it
impossible for files to be created in the sites/default/files directory
before we create the symlink for that directory.