Add option to compress old releases
We (District09) were having some issues where our inodes usage was
getting out of hand (90%). Since a lot of composer-based PHP projects
use a lot of files (in the vendor folder), a lot of inodes are being
used. Since we keep old releases (the amount of which can be
configured), we use even more inodes. This commit adds the option to
compress said old releases, so that we reduce the amount of used inodes
to one per old release. This _should_ fix this particular problem.