Update code to be more "DRY"
In stead of trying to get _all_ previous releases, figuring out which
one is the new one so we don't compress it and then compressing all the
others, only get the last release (before the one we're deploying right
now) and only compress that one. Eventually (after a couple of releases)
all previous releases will be compressed. It makes more sense that this
setting only has effect on releases once it has been changed, and leaves
the older releases alone. It also make the code a bit less complicated,
and makes it so we can reuse some code.