Refactor the whole package to use events
Because of https://github.com/consolidation/annotated-command/pull/273,
which technically speaking contains a BC break, it was made clear for us
that we're not using Robo or the consolidation packages as they were
intended. In an effort to keep this package as extendible as possible,
and keeping the same philosophy as before, we switched to an event-based
system. Basically every step in a command fires an event. The event
listener returns a `HandlerWithPriority`, which it turn returns a
`TaskInterface` to be executed for this step in the command. We made our
own implementation of event priorities, since the default implementation
used in Robo doesn't support them
(https://github.com/consolidation/annotated-command/issues/244).

I tried to document everything as well as possible in the README, but it
might (probably will) need some lovin'.

This package contains default event handlers with some sensible (to us)
defaults. You can prevent these handlers from being executed by writig
your own handler with a higher priority (lower number) which does its
thing and then calls `$event->stopPropagation()`.